- Adicionado lógica para verificar se é necessário aplicar os juros no valor total ou não; - Adicionado lógica de cálculo dos juros, com base na taxa Selic atual, sobre o valor total da negociação; - Finalizado implementação do service para o endpoint /negociacao; - Adicionado documentação do service SelicService

+++ src/main/java/com/github/douglasmartins/calculajuros/model/PaymentInstallment.java
@@ -13,7 +13,7 @@ public class PaymentInstallment {
 
     private Integer numeroParcela;
 
-    private BigDecimal valor;
+    private Double valor;
 
     private BigDecimal taxaJurosAoMes;
 }
